Feature ETP-1058: Implement displayLogic and readOnlyLogic.
