+++ .github/workflows/docker-build-and-publish.yml
@@ -1,4 +1,4 @@
-name: Build and Publish Docker Image on PR
+name: Docker Build and Publish
 
 on:
   push:
