+++ packages/ComponentLibrary/.eslintrc.cjs
@@ -14,5 +14,6 @@ module.exports = {
       'warn',
       { allowConstantExport: true },
     ],
+    'react/display-name': 'warn'
   },
 }
