--- packages/ComponentLibrary/src/components/FormView/selectors/StringSelector.tsx
@@ -1,3 +1,0 @@
-import TextInputBase from '../../Input/TextInput/TextInputBase';
-
-export const StringSelector = TextInputBase;
