+++ packages/MainUI/.eslintrc.json
@@ -22,7 +22,7 @@
   },
   "rules": {
     "react/react-in-jsx-scope": "off",
-    // "react/prop-types": "off",
+    "react/prop-types": "off",
     "@typescript-eslint/no-unused-vars": [
       "warn",
       {
@@ -32,6 +32,7 @@
     ],
     "@typescript-eslint/explicit-function-return-type": "off",
     "@typescript-eslint/explicit-module-boundary-types": "off",
-    "no-console": "off"
+    "no-console": "off",
+    "react/display-name": "warn"
   }
 }
