+++ packages/MainUI/components/Form/FormView/selectors/TableDirSelector.tsx
@@ -1,48 +1,45 @@
 import { useCallback, useMemo, useEffect, useState } from 'react';
-import SearchOutlined from '../../../assets/icons/search.svg';
-import { useComboSelect } from '@workspaceui/etendohookbinder/src/hooks/useComboSelect';
+import SearchOutlined from '@workspaceui/componentlibrary/src/assets/icons/search.svg';
+import { useDatasource } from '@workspaceui/etendohookbinder/src/hooks/useDatasource';
 import { useTheme } from '@mui/material';
-import { Option, SearchSelectorProps } from '../types';
-import Spinner from '../../Spinner';
-import Select from '../../Input/Select';
+import { Option, TableDirSelectorProps } from '../types';
+import Spinner from '@workspaceui/componentlibrary/src/components/Spinner';
+import Select from '@workspaceui/componentlibrary/src/components/Input/Select';
 
 const getOptionLabel = (option: Option) => option.title;
 
 const optionEqualValue = (option: Option, value: { id: string }) => option.id === value.id || option.value === value.id;
 
-const SearchSelector = ({ onChange, value, field, name, disabled, readOnly }: SearchSelectorProps) => {
+const TableDirSelector = ({ onChange, entity, value, name, disabled, readOnly }: TableDirSelectorProps) => {
   const theme = useTheme();
-  const { records, loading, error, loaded } = useComboSelect(field);
+  const { records, loading, error, loaded } = useDatasource(entity);
   const [selectedValue, setSelectedValue] = useState<Option | null>(null);
 
   const isDisabled = disabled || readOnly;
 
-  const options = useMemo(() => {
-    const valueField = (field.original.selector?.valueField ?? '') as string;
-
-    return records.map(record => ({
-      id: record[valueField] as string,
-      title: (record._identifier || record.name || record.id) as string,
-      value: record[valueField] as string,
-    }));
-  }, [field.original?.selector?.valueField, records]);
+  const options = useMemo(
+    () =>
+      records.map(record => ({
+        id: record.id as string,
+        title: (record._identifier || record.name || record.id) as string,
+        value: record.id as string,
+      })),
+    [records],
+  );
 
   useEffect(() => {
+    let option;
+
     if (value && options.length > 0) {
-      const option = options.find(opt => {
+      option = options.find(opt => {
         if (typeof value === 'object' && 'id' in value) {
           return opt.id === value.id || opt.value === value.id;
         }
         return opt.id === String(value) || opt.value === String(value);
       });
-      if (option) {
-        setSelectedValue(option);
-      } else {
-        setSelectedValue(null);
-      }
-    } else {
-      setSelectedValue(null);
     }
+
+    setSelectedValue(option ?? null);
   }, [value, options]);
 
   const handleChange = useCallback(
@@ -72,4 +69,4 @@ const SearchSelector = ({ onChange, value, field, name, disabled, readOnly }: Se
   );
 };
 
-export default SearchSelector;
+export default TableDirSelector;
