--- packages/MainUI/screens/Form/types.ts
@@ -1,1 +1,0 @@
-export * from '@workspaceui/componentlibrary/src/components/FormView/types';
